+++ app/code/core/Mage/OAuth/Model/Resource/Token.php
@@ -73,19 +73,23 @@ class Mage_OAuth_Model_Resource_Token extends Mage_Core_Model_Resource_Db_Abstra
     /**
      * Delete old entries
      *
-     * @param $minutes
+     * @param int $minutes
      * @return int
      */
     public function deleteOldEntries($minutes)
     {
-        $adapter = $this->_getWriteAdapter();
-        return $adapter->delete(
-            $this->getTable('oauth/token'),
-            $adapter->quoteInto(
-                'created_at <= TIMESTAMPADD(MINUTE, -?, CURRENT_TIMESTAMP())',
-                $minutes,
-                Zend_Db::INT_TYPE
-            )
-        );
+        if ($minutes > 0) {
+            $adapter = $this->_getWriteAdapter();
+
+            return $adapter->delete(
+                $this->getMainTable(),
+                $adapter->quoteInto(
+                    'type = "' . Mage_OAuth_Model_Token::TYPE_REQUEST . '" AND created_at <= ?',
+                    Varien_Date::formatDate(time() - $minutes * 60)
+                )
+            );
+        } else {
+            return 0;
+        }
     }
 }
